converter-and-renderer/requirements.txt
numpy
moderngl
pillow
pyrr
PyQt5==5.9.1
matplotlib
